Add dialog

+++ src/app/core/components/header/header.component.ts
@@ -5,6 +5,8 @@ import { MatMenuModule } from '@angular/material/menu';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { Router } from '@angular/router';
 import { TokenService } from '../../services/token.service';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogComponent } from '../dialog/dialog.component';
 
 @Component({
   selector: 'app-header',
@@ -15,6 +17,7 @@ import { TokenService } from '../../services/token.service';
 export class HeaderComponent {
   private router = inject(Router);
   private tokenService = inject(TokenService);
+  private dialog = inject(MatDialog);
 
   userRole = this.tokenService.getDecodedPayload() ? this.tokenService.getDecodedPayload().ROLE : 'comum';
 
@@ -23,8 +26,20 @@ export class HeaderComponent {
   }
 
   logout() {
-    this.tokenService.removeToken();
-    this.router.navigate(['/login']);
+    const dialogRef = this.dialog.open(DialogComponent, {
+      enterAnimationDuration: '250ms',
+      data: {
+        titulo: 'Confirmação de Logout',
+        conteudo: 'Tem certeza que deseja sair do sistema?'
+      }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.tokenService.removeToken();
+        this.router.navigate(['/login']);
+      }
+    });
   }
 
   goToConfigs() {
